Major (>10%) performance improvement in BSP rendering.
This was accomplished with two related optimizations. 

First, the renderer now keeps track of whether a surface's lightmap texture is
the same as its last one, in addition to its regular texture. This was a
golden opportunity, since an entire map might have just two or three lightmap
textures, so the proportion of lightmap texture switches which were redundant
was very high. 

Second, the renderer wasn't fully utilizing the fact that it already tracked
texture changes, because certain things (like the surface blood effects,
parallax maps, alpha test, etc.) are tied to the texinfo struct, therefore the
corresponding OpenGL calls only need to be made if the texinfo is different 
from the last one.

I suspect there's more room for improvement though. I'll bet if we grouped as
many surfaces from each leaf into as few VBOs as possible (all quads with the 
same texture and same lightmap texture, then another VBO for the tris, etc.) 
we could get another 10% boost at least. Preliminary calculations indicate 
that on zorn2k11, the number of VBOs and thus the number of glDrawArrays calls
would be reduced by around 18%.