Smooth out movement in the hovercraft. Add attribution to g_vehicles.c.
The bug that caused the weird hovercraft movement was in the original jetpack
tutorial, justified as preventing the player from going too fast in vehicles,
however I checked with the speedometer and removing that code did not increase
speed in any of the vehicles. 

Since I'd already found the original tutorial, I added a credit to its author
and a link at the beginning of g_vehicles.c.