Whoops, the previous commit message should have been the following:
MD2 lerping now handled in GLSL for shell and regular mesh rendering.
Mirror/glass rendering still uses the old system for now.
The GPUANIM uniform in the shader should be set to 2 to enable MD2 lerping.
Because MD2s with lots of frames require so many separate VBOs, the maximum
number of VBO vertex caches has been increased, and vertStoreMod_t has been
changed to allow the data for an individual frame to be retrieved.

(Blank commit just to get the correct description in the commit logs.)