Add cvar log_dest_udp for remote server logging/monitoring.
Set it to an ip:port to echo the console to that IP using connectionless UDP
"print" packets. Still no batching, buffering, or rate limiting, so it's
possible to lose data if a lot of text is printed at once (e.g. cvarlist.)
However, for normal gameplay it should be fine, and for once-off commands with
a lot of output, we have rcon. 

Idea from Xonotic, but this implementation is much simpler.

One possible use is two-way bridge between ingame chat and IRC.