Optimize CM_FastTrace for BSPs without much detail.
This is accomplished by, for each BSP node, tracking a box volume which is
completely empty of brushes. If a trace lies completely inside this volume,
it is unblocked by BSP geometry.

There may be a slight slowdown for maps that have a lot of detailed BSP
geometry, but not severe.

This optimization is helpful for terrain maps, where there is very little BSP
detail.