menu.c: fix GetLine function not terminating copy of final line.
The GetLine function is supposed to return a copy of the first line in an
input string, but if the input has only one line in it, it would fail to
terminate the string, leading to possible buffer overruns.

This bug probably explains the crashes people were seeing in the "host server"
menu.