menu: GetLine function now returns static string.
Previously it would allocate a new string which would have to be freed.
However, we don't currently need that feature anywhere, so it just returns a
static string. This means that calling GetLine twice will invalidate its first
return value.

This commit also fixes a memory leak in the "host server" submenu.