client: prediction: unbreak up stair smoothing with cl_test enabled.
The stair smoothing is a special feature that allows some interpolation
of the z-position of the camera while walking up and down stairs. When I
added the cl_test code to the client-side prediction, I failed to adjust
the stair smoothing code to work well with it, resulting in jumpiness as the
stair smoothing code had outdated information.