ref_gl: r_bloom: experimental GLSL shader for darkening passes. r_test to use.
Results should be visually identical.

Thus far I can't find a real performance difference (not surprising since the
bloom scratch texture is so small,) but it reduces the number of draw calls,
and eliminates another use of the fixed-function pipeline. 

Soon I'm going to be making the bloom scratch texture much larger, we'll see
if I can pick out a performance difference then.