qcommon: cmodel: Add APIs for exporting all terrain collision geometry.
Will be used by other parts of the code that might need this geometry, such
as the ragdoll system. The geometry is all treated as if it is a single mesh.
There are APIs for getting the number of verts and the number of tris, for
getting a single vert as a vector, and for getting a single triangle as an
array of three integer indices.