add "downloadhttp" command for downloading files from the HTTP repository.
You need to know the file's path (i.e. "maps/dm-titan.bsp") and it won't
autodownload accessory files such as .fog scripts, but this is still useful
for downloading files from the ingame console.

Also revamp UDP "download" command to not attempt a download if not connected
to a server. This avoids an assertion failure.