god rays: use depth checking for BSP/terrain occlusion instead of redrawing.
We compute a bounding rectangle for the sun sprite, and then only blit that
small part of the depth buffer over to the RBO. Then, we render the sun sprite
with depth checking.

Before, we were actually re-rendering BSP geometry off screen, and we weren't
handling terrain meshes at all.

This gives a nice performance gain.