qrad3 (terrain): prepare for multiple-channel lightmaps.
Direct non-sun light, direct sunlight, and indirect light are now computed
separately and (for now) combined at the end. Verified that this makes no
visual difference to the current output.

By unifying some code, this commit also enables -ambient, -maxlight, -scale,
and -grayscale on terrain lightmaps, and it also makes eventual support for
radiosity easier to implement on terrain.

I measured an 18% performance hit with this change on terraintest with -refine
and -terrainrefine both set to 8.